src/main/java/ds/algos/linkedList/ReverseLinkedList.java
package ds.algos.linkedList;

public class ReverseLinkedList {
    public ListNode reverseList(ListNode head) {
        ListNode prev = null;
        ListNode current = head;
        while (current != null) {
            ListNode next = current.next;
            current.next = prev;
            prev = current;
            current = next;
        }
        return prev;
    }

    public ListNode reverseList_practice(ListNode head) {
        ListNode current = head;
        ListNode prev = null;
        while (current != null) {

            ListNode next = current.next;

            current.next = prev;

            prev = current;

            current = next;
        }
        return prev;
    }




















    }
